requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.8
psycopg==3.2.12
psycopg-binary==3.2.12
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
django-crispy-forms==2.5
crispy-bootstrap5==0.7
